refactor: change helpers structure, general optimizations
