+++ src/components/desktop/charts/CryptoChart.vue
@@ -43,8 +43,8 @@ import ChartOptions from "D#/charts/ChartOptions.vue";
 import CandlestickChart from "D#/charts/CandlestickChart.vue";
 import { TIMES } from "@/config/config.js";
 import { computed, ref, toRef, watch } from "vue";
-import { formatValue, formatPercentage } from "@/helpers/formatNumber.helper";
-import { calculatePercentage } from "@/helpers/getPrice.helper";
+import { formatValue, formatPercentage } from "@/helpers/formatter.helper";
+import { calculatePercentage } from "@/helpers/crypto.helper";
 import { useCryptoStore } from "S#/crypto.store";
 import Big from "@/helpers/big.helper";
