+++ src/components/desktop/dashboard/AssetAnalysis.vue
@@ -2,7 +2,7 @@
   <tr class="h4 bg-dark transition" @click="open = !open">
     <td class="align-center" id="chart">
       <img 
-        :src="getCryptoIcon(crypto)" 
+        :src="getIcon(crypto)" 
         :alt="crypto"
         onerror="this.src='/src/assets/icons/generic.svg'"
       >
@@ -35,14 +35,12 @@
 <script setup>
 import LineChart from "D#/charts/LineChart.vue";
 import TransactionsAnalysis from "D#/dashboard/TransactionsAnalysis.vue";
-import { onMounted, ref, toRefs, watch } from "vue";
-import getCryptoIcon from "@/helpers/getCryptoIcon.helper.js";
+import { onMounted, ref, toRefs, watch, computed } from "vue";
 import Big from "@/helpers/big.helper.js";
 import { useCryptoStore } from "S#/crypto.store";
-import { formatValue, formatPercentage } from "@/helpers/formatNumber.helper";
-import { computed } from "@vue/reactivity";
-import { getDollarPrice } from "@/helpers/getPrice.helper";
-import { addEarnings } from "@/helpers/getPastQuantity.helper";
+import { formatValue, formatPercentage } from "@/helpers/formatter.helper";
+import { getDollarPrice, getIcon } from "@/helpers/crypto.helper";
+import { addEarnings } from "@/helpers/transactions.helper";
 import { ANALYSIS_TIMES } from "@/config/config";
 
 const props = defineProps({
