+++ src/components/desktop/dashboard/TransactionAnalysis.vue
@@ -6,14 +6,13 @@
     <td>{{ formatValue(transaction.price) }}</td>
     <td :class="parseFloat(transaction.earnings) > 0 ? 'green' : parseFloat(transaction.earnings) < 0 ? 'red' : ''">{{ formatValue(transaction.earnings) }}</td>
     <td :class="parseFloat(transaction.change) > 0 ? 'green' : parseFloat(transaction.change) < 0 ? 'red' : ''">{{ formatPercentage(transaction.change) }}</td>
-    <td>{{ dateFormat(transaction.date) }}</td>
+    <td>{{ formatDate(transaction.date) }}</td>
   </tr>
 </template>
 
 <script setup>
 import Big from "@/helpers/big.helper.js";
-import { formatValue, formatPercentage } from "@/helpers/formatNumber.helper.js";
-import { dateFormat } from "@/helpers/formatDate.helper.js";
+import { formatValue, formatPercentage, formatDate } from "@/helpers/formatter.helper.js";
 
 defineProps({
   crypto: {
