+++ src/components/desktop/markets/Card.vue
@@ -3,12 +3,12 @@
     <td>
       <div class="align-center">
         <img 
-          :src="iconUrl" 
+          :src="getIcon(crypto)" 
           :alt="crypto" 
           onerror="this.onerror = null; this.src='/src/assets/icons/generic.svg'"
           class="icon"
         >
-        <span class="title">{{ name }}</span>
+        <span class="title">{{ store.tickerInfo[crypto].name || crypto }}</span>
       </div>
     </td>
 
@@ -43,9 +43,8 @@ import Icon from "U#/Icon.vue";
 import Button from "U#/Button.vue";
 import { computed, ref, watch } from "vue";
 import { useCryptoStore } from "S#/crypto.store";
-import { getDollarPrice, getPercentageChange } from "@/helpers/getPrice.helper";
-import { formatPercentage, formatValue } from "@/helpers/formatNumber.helper";
-import getCryptoIcon from "@/helpers/getCryptoIcon.helper";
+import { getDollarPrice, getPercentageChange, getIcon } from "@/helpers/crypto.helper";
+import { formatPercentage, formatValue } from "@/helpers/formatter.helper";
 
 const { crypto } = defineProps({
   crypto: {
@@ -54,10 +53,7 @@ const { crypto } = defineProps({
   }
 });
 
-const iconUrl = getCryptoIcon(crypto);
-
 const store = useCryptoStore();
-const name = store.tickerInfo[crypto].name || crypto;
 
 const price = computed(() => formatValue(getDollarPrice(crypto, store.prices)));
 const mcap = formatValue(store.tickerInfo[crypto].mcap);
