+++ src/components/desktop/wallet/Transaction.vue
@@ -2,7 +2,7 @@
   <tr :class="'transition nohover ' + fontSize + ' ' + (shorter ? 'short' : '')">
     <td v-if="withTicker" class="align-center">
       <img 
-        :src="getCryptoIcon(crypto)" 
+        :src="getIcon(crypto)" 
         :alt="crypto"
         onerror="this.onerror = null; this.src='/src/assets/icons/generic.svg'"
       >
@@ -12,9 +12,9 @@
     <td>{{ crypto }}/{{ base }}</td>
     <td>{{ isBuy ? 'BUY' : 'SELL' }}</td>
     <td>{{ quantity }}</td>
-    <td>{{ price }}</td>
+    <td>{{ Big(price).toFormat(2) }}</td>
     <td :class="isHigher ? 'green' : isHigher !== null ? 'red' : ''">${{ value && value.toFormat(2) || "" }}</td>
-    <td>{{ date }}</td>
+    <td>{{ formatDate(date) }}</td>
     <td v-if="!withTicker && $route.params.isAuth">
       <Button btnClass="bg-outline h4" @click="displayPopup = true">Details</Button>
     </td>
@@ -32,11 +32,11 @@
 import Button from "U#/Button.vue";
 import TransactionPopup from "D#/wallet/TransactionPopup.vue";
 import { ref, computed, watch } from "vue";
-import { dateFormat } from "@/helpers/formatDate.helper";
-import { getDollarPrice } from "@/helpers/getPrice.helper";
+
+import { formatDate } from "@/helpers/formatter.helper";
+import { getDollarPrice, getIcon } from "@/helpers/crypto.helper";
 import Big from "@/helpers/big.helper";
 import { useCryptoStore } from "S#/crypto.store";
-import getCryptoIcon from "@/helpers/getCryptoIcon.helper";
 
 const { transaction , crypto } = defineProps({
   transaction: {
@@ -67,15 +67,12 @@ const { transaction , crypto } = defineProps({
 
 const store = useCryptoStore();
 let { isBuy, base, quantity, price, date } = transaction;
-date = dateFormat(date);
 
 let value;
 if (isBuy) 
   value = computed(() => Big(getDollarPrice(crypto, store.prices)).times(quantity));
 else
   value = Big(price).times(quantity);
-
-price = new Big(price).toFormat(2);
 
 const displayPopup = ref(false);
