+++ src/components/desktop/wallet/Wallet.vue
@@ -30,10 +30,11 @@
 import WalletRow from "D#/wallet/WalletRow.vue";
 import Popup from "U#/Popup.vue";
 import byMcap from "@/helpers/sortByMcap.helper.js";
+import { computed, ref, toRefs } from "vue";
+
 import { QUOTES_DOLLAR } from "@/config/config.js";
 import { useAuthStore } from "S#/auth.store.js";
 import { useCryptoStore } from "S#/crypto.store.js";
-import { computed, ref, toRefs } from "vue";
 
 const props = defineProps({
   search: {
