+++ src/components/mobile/wallet/Transaction.mobile.vue
@@ -2,7 +2,7 @@
   <tr :class="fontSize">
     <td v-if="withTicker" class="align-center">
       <img 
-        :src="getCryptoIcon(crypto)" 
+        :src="getIcon(crypto)" 
         :alt="crypto"
         onerror="this.onerror = null; this.src='/src/assets/icons/generic.svg'"
       >
@@ -10,7 +10,7 @@
     </td>
     <td>{{ isBuy ? 'BUY' : 'SELL' }}</td>
     <td>{{ quantity }}</td>
-    <td>{{ price }}</td>
+    <td>{{ Big(price).toFormat(2) }}</td>
     <td v-if="!withTicker && $route.params.isAuth" class="right-align">
       <Button btnClass="bg-outline h4" btnCss="padding: 0.3em;" @click="displayPopup = true">Details</Button>
     </td>
@@ -28,11 +28,10 @@
 import Button from "U#/Button.vue";
 import TransactionPopup from "M#/wallet/TransactionPopup.mobile.vue";
 import { ref, computed, watch } from "vue";
-import { dateFormat } from "@/helpers/formatDate.helper";
-import { getDollarPrice } from "@/helpers/getPrice.helper";
+import { getDollarPrice } from "@/helpers/crypto.helper";
 import Big from "@/helpers/big.helper";
 import { useCryptoStore } from "S#/crypto.store";
-import getCryptoIcon from "@/helpers/getCryptoIcon.helper";
+import { getIcon } from "@/helpers/crypto.helper";;
 
 const { transaction , crypto } = defineProps({
   transaction: {
@@ -62,16 +61,13 @@ const { transaction , crypto } = defineProps({
 });
 
 const store = useCryptoStore();
-let { isBuy, base, quantity, price, date } = transaction;
-date = dateFormat(date);
+let { isBuy, base, quantity, price } = transaction;
 
 let value;
 if (isBuy) 
   value = computed(() => Big(getDollarPrice(crypto, store.prices)).times(quantity));
 else
   value = Big(price).times(quantity);
-
-price = new Big(price).toFormat(2);
 
 const displayPopup = ref(false);
