+++ src/config/charts/candlestickChartOptions.js
@@ -1,6 +1,5 @@
 import extractSass from "@/helpers/extractSass.helper.js";
-import { timeFormat, dateFormat } from "@/helpers/formatDate.helper.js";
-import { formatValue } from "@/helpers/formatNumber.helper.js";
+import { formatValue, formatDate, formatTime } from "@/helpers/formatter.helper.js";
 
 const gridColor = extractSass("grid-color");
 const textColor = extractSass("primary");
@@ -22,7 +21,7 @@ export default {
   xAxis: { 
     type: "time",
     axisLabel: {
-      formatter: timeFormat,
+      formatter: formatTime,
       color: textColor,
       fontSize: "1em"
     },
@@ -36,7 +35,7 @@ export default {
 
     axisPointer: {
       label: {
-        formatter: ({ value }) => dateFormat(value),
+        formatter: ({ value }) => formatDate(value),
       }
     },
