--- src/helpers/getBaseLCM.helper.js
@@ -1,10 +1,0 @@
-export default function(cryptos, tickers) {
-  if (!cryptos.length || !Object.keys(tickers).length) return [];
-  let ret = tickers[cryptos[0]].quotes;
-
-  for (let i = 1; i < cryptos.length; i++) {
-    ret = ret.filter(value => tickers[cryptos[i]].quotes.includes(value));
-  };
-
-  return ret;
-}
