+++ src/stores/crypto.store.js
@@ -1,8 +1,8 @@
 import { defineStore } from "pinia";
 import { fetchBinance, fetchServer } from "@/helpers/fetch.helper.js";
 import createSocket from "@/helpers/createSocket.helper";
-import { getPastQuantity } from "@/helpers/getPastQuantity.helper";
-import { getPair } from "@/helpers/getPrice.helper";
+import { getPastQuantity } from "@/helpers/transactions.helper";
+import { getPair } from "@/helpers/crypto.helper";
 import { KLINES_LIMIT } from "@/config/config";
 import { useAuthStore } from "S#/auth.store";
