+++ src/views/mobile/Dashboard.mobileView.vue
@@ -40,7 +40,7 @@ import { useAuthStore } from "S#/auth.store";
 import { useCryptoStore } from "S#/crypto.store";
 import Big from "@/helpers/big.helper";
 import { computed, ref } from "vue";
-import { getDollarPrice } from "@/helpers/getPrice.helper";
+import { getDollarPrice } from "@/helpers/crypto.helper";
 
 const authStore = useAuthStore();
 const cryptoStore = useCryptoStore();
